[Refactor] useDrawOkmok 훅 구현
