+++ hybrid/src/components/game/OmokBoard.tsx
@@ -1,14 +1,14 @@
 import { useEffect, useRef, useState } from 'react';
 import styled from 'styled-components';
 import TimerAndPut from '@/components/game/TimerAndPut';
-import { useRecoilValue } from 'recoil';
-import { putInfoState } from '@/utils/recoil/socket';
 import PointBoard from './PointBoard';
+import useDrawOkmok from '@/hooks/useDrawOkmok';
 
 export default function OmokBoard() {
   const canvasRef = useRef<HTMLCanvasElement>(null);
+  const canvas = canvasRef.current;
+  useDrawOkmok(canvas);
   const [pointXY, setPointXY] = useState<number[]>([-1, -1]);
-  const putInfo = useRecoilValue(putInfoState);
   const pointXYHandler = (x: number, y: number) => {
     setPointXY([x, y]);
   };
@@ -34,20 +34,6 @@ export default function OmokBoard() {
     }
   };
 
-  const drawOkmok = (x: number, y: number, color: string) => {
-    if (x < 0 || y < 0) return;
-    const canvas = canvasRef.current;
-    if (!canvas) return;
-    const ctx = canvas.getContext('2d');
-    if (!ctx) return;
-    ctx.beginPath();
-    ctx.strokeStyle = color;
-    ctx.fillStyle = color;
-    ctx.arc(blank + x * 25, blank + y * 25, 10, 0, 2 * Math.PI);
-    ctx.fill();
-    ctx.stroke();
-  };
-
   useEffect(() => {
     if (!canvasRef) return;
     const canvas = canvasRef.current;
@@ -56,12 +42,6 @@ export default function OmokBoard() {
     canvas.setAttribute('height', '375');
     drawBoard(canvas);
   }, []);
-
-  useEffect(() => {
-    const { x, y, player } = putInfo;
-    const color = player === 1 ? '#00ff00' : '#ff00ff';
-    drawOkmok(x, y, color);
-  }, [putInfo]);
 
   return (
     <div>
